modified:   src/components/App/App.tsx modified:   src/components/Auth/firebase.ts modified:   src/components/Conversations/Conversation.tsx modified:   src/components/Conversations/InitialMenu.tsx modified:   src/components/Conversations/MessageInput.tsx deleted:    src/components/Conversations/MessageList.tsx modified:   src/components/Conversations/SettingsModal.tsx modified:   src/components/Conversations/Sidebar.tsx modified:   src/components/Conversations/TokenCounter.tsx modified:   src/components/Conversations/types/Conversations.types.ts modified:   src/hooks/useAIResponse.ts modified:   src/hooks/useEditing.ts modified:   src/hooks/useSidebar.ts modified:   src/index.tsx modified:   src/utils/openAIUtil.ts

+++ src/components/App/App.tsx
@@ -1,7 +1,8 @@
+//App.tsx
 import React, { useState, useContext, useEffect, useRef } from "react";
 import { AuthContext } from '../Auth/AuthContext';
 import { ConversationType, SystemPromptType } from '../Conversations/types/Conversations.types';
-import { fetchConversations, updateConversations, deleteConversation } from '../Auth/firebase';
+import { fetchUserData, updateConversations, deleteConversation } from '../Auth/firebase';
 import { ScrollWrapper, MainContainer, SidebarContainer, Placeholder } from './App.styles'
 import Topbar from '../Conversations/Topbar'
 import Sidebar from '../Conversations/Sidebar'
@@ -12,49 +13,59 @@ import { Spinner } from "../Conversations/Spinner";
 const App: React.FC = () => {
   const { user } = useContext(AuthContext);
   const [conversations, setConversations] = useState<ConversationType[]>([]);
-  const [activeConversation, setActiveConversation] = useState<ConversationType | null>(null);
   const [systemprompts, setSystemPrompts] = useState<SystemPromptType[]>([]);
+  const [activeConversation, setActiveConversation] = useState<ConversationType | null>(null);
+  const [receivingId, setReceivingId] = useState('');
   const [showMenu, setShowMenu] = useState(true);
   const [model, setModel] = useState('gpt-3.5-turbo-0613'); 
   const [apiKey, setApiKey] = useState('');
-  const [receivingId, setReceivingId] = useState('');
   const [sidebarWidth, setSidebarWidth] = useState(200);
   const [sidebarTransition, setSidebarTransition] = useState(false)
   const minSidebarWidth = 15;
   const maxSidebarWidth = 600;
   const scrollWrapperRef = useRef(null);
 
+  const handleUpdateConversations = async (updatedConversation: ConversationType, shouldUpdateFirestore: boolean = true) => {
+    const updatedConversations = conversations.map(Conversation => 
+      Conversation.id === updatedConversation.id ? updatedConversation : Conversation
+    );
+    setConversations(updatedConversations);
+    
+    if (shouldUpdateFirestore) {
+      await updateConversations(user?.uid, updatedConversations);
+    }
+  };
+
+  useEffect(() => {
+    const updatedActiveConversation = conversations.find(conv => conv.id === activeConversation?.id);
+
+    if (updatedActiveConversation) {
+      setActiveConversation(updatedActiveConversation);
+    } else if (!conversations.some(conv => conv.id === activeConversation?.id)) {
+      setActiveConversation(null);
+    }
+  }, [conversations]);
+
+  const handleDeleteConversation = async (id: string) => {
+    if (window.confirm('Are You Sure to Delete?')) {
+      await deleteConversation(user?.uid, id);
+      setConversations(conversations.filter(conv => conv.id !== id));
+      }
+  };
+
   const handleResize = (width: number) => {
     setSidebarWidth(width);
   };
 
   useEffect(() => {
-    const fetchUserConversations = async () => {
-      const result = await fetchConversations(user?.uid);
-      if (Array.isArray(result)) return;
-      const { messages, systemPrompts } = result;
-      const conversations: ConversationType[] = messages.map((message: any) => ({id: message.id, ...message} as ConversationType));
-      const systemprompts: SystemPromptType[] = systemPrompts.map((systemPrompt: any) => ({id: systemPrompt.id, ...systemPrompt} as SystemPromptType));
-      setConversations(conversations);
-      setSystemPrompts(systemprompts);
+    const getUserData = async () => {
+      const fetchedData = await fetchUserData(user?.uid);
+      setConversations(fetchedData.conversations);
+      setSystemPrompts(fetchedData.systemPrompts);
     };
-
-    fetchUserConversations();
+  
+    getUserData();
   }, [user?.uid]);
-
-  const handleUpdateConversations = async (updatedConversation: ConversationType) => {
-    const updatedConversations = conversations.map(item => 
-      item.id === updatedConversation.id ? updatedConversation : item
-    );
-
-    await updateConversations(user?.uid, updatedConversations);
-    setConversations(updatedConversations);
-  };
-
-  const handleDeleteConversation = async (id: string) => {
-    await deleteConversation(user?.uid, id);
-    setConversations(conversations.filter(conv => conv.id !== id));
-  };
 
   if (!user) {
     return <div style={{padding: '30px'}}><Spinner/></div>;
@@ -75,6 +86,7 @@ const App: React.FC = () => {
         systemprompts={systemprompts}
         setSystemPrompts={setSystemPrompts}
         setSidebarTransition={setSidebarTransition}
+        
       />
       <MainContainer>
         <SidebarContainer showMenu={showMenu} sidebarWidth={sidebarWidth} sidebarTransition={sidebarTransition} tabIndex={0}>
@@ -84,7 +96,8 @@ const App: React.FC = () => {
             setConversations={setConversations}
             setActiveConversation={setActiveConversation}
             handleUpdateConversations={handleUpdateConversations}
-            deleteConversation={handleDeleteConversation}
+            handleDeleteConversation={handleDeleteConversation}
+            receivingId={receivingId}
           />
         </SidebarContainer>
         {showMenu && (
@@ -92,7 +105,7 @@ const App: React.FC = () => {
         )}
         {activeConversation ? (
           <Conversation
-            conversation={activeConversation}
+            activeConversation={activeConversation}
             model={model}
             apiKey={apiKey}
             handleUpdateConversations={handleUpdateConversations}
